delete pathbaseexecutor
