+++ src/graph/executor/CMakeLists.txt
@@ -6,7 +6,6 @@ nebula_add_library(
     executor_obj OBJECT
     Executor.cpp
     StorageAccessExecutor.cpp
-    PathBaseExecutor.cpp
     logic/LoopExecutor.cpp
     logic/PassThroughExecutor.cpp
     logic/StartExecutor.cpp
