+++ src/graph/executor/StorageAccessExecutor.h
@@ -11,6 +11,13 @@
 #include "graph/executor/Executor.h"
 #include "graph/util/Utils.h"
 
+using nebula::storage::StorageRpcResponse;
+using nebula::storage::cpp2::GetNeighborsResponse;
+using RpcResponse = StorageRpcResponse<GetNeighborsResponse>;
+using PropRpcResponse = StorageRpcResponse<nebula::storage::cpp2::GetPropResponse>;
+using VertexProp = nebula::storage::cpp2::VertexProp;
+using nebula::storage::StorageClient;
+
 namespace nebula {
 
 class Expression;
@@ -167,6 +174,18 @@ class StorageAccessExecutor : public Executor {
                                                          Expression *expr,
                                                          bool dedup,
                                                          bool isCypher = false);
+
+  bool hasSameEdge(const std::vector<Value> &edgeList, const Edge &edge);
+
+  void addGetNeighborStats(RpcResponse &resp, size_t stepNum, bool reverse);
+
+  folly::Future<std::vector<Value>> getProps(const std::vector<Value> &vids,
+                                             const std::vector<VertexProp> *vertexPropPtr);
+
+  std::vector<Value> handlePropResp(PropRpcResponse &&resps);
+
+ protected:
+  folly::SpinLock statsLock_;
 };
 
 }  // namespace graph
